PC New Movement
PC now 'wanders' to a random waypoint when it is not near an enemy/resource. Added a function that gets the list of the nearest x amount of waypoints and returns a list of them. The PC then randomly 'selects' one of them and travels to it in search of enemies/resources.